Fix context overflow classification and wiring
